rename: mirror to absolute
